update imports for FancyMenu
