+++ modules/food-menu/index.js
@@ -2,4 +2,14 @@
 
 export {FancyMenu as FoodMenu} from './fancy-menu'
 
-export type {MenuItemType, StationMenuType, CorIconType, MenuItemContainerType, ItemCorIconMapType, MasterCorIconMapType, ProcessedMealType} from './types'
+export type {
+	MenuItemType,
+	StationMenuType,
+	CorIconType,
+	MenuItemContainerType,
+	ItemCorIconMapType,
+	MasterCorIconMapType,
+	ProcessedMealType,
+	DayPartsCollectionType,
+	DayPartMenuType,
+} from './types'
