+++ modules/food-menu/lib/choose-meal.js
@@ -1,9 +1,9 @@
 // @flow
 
 import type momentT from 'moment'
-import type {ProcessedMealType} from '../../../source/views/menus/types'
+import type {ProcessedMealType} from '../types'
 import type {FilterType} from '@frogpond/filter'
-import {findMeal} from '../../../source/views/menus/lib/find-menu'
+import {findMeal} from './find-menu'
 
 export function chooseMeal(
 	meals: ProcessedMealType[],
