+++ source/views/menus/lib/__tests__/find-menu.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 // @flow
-import {findMenu} from '../find-menu'
+import {findMenu} from '../../../../../modules/food-menu/lib/find-menu'
 import moment from 'moment-timezone'
 const CENTRAL_TZ = 'America/Winnipeg'
 import type {DayPartsCollectionType} from '../../types'
